iox-#859 Move typed allocator to iox::memory
